+++ site/js/ponymail.js
@@ -1950,6 +1950,31 @@ function showSpinner(show) {
 }
 
 
+// Ephemeral configuration - non-account but still saved through reloads
+
+// Saving prefs as a json string
+function saveEphemeral() {
+    if (typeof(window.localStorage) !== "undefined") {
+        window.localStorage.setItem("ponymail_config_ephemeral", JSON.stringify(prefs))
+    }
+}
+
+// load ephemeral prefs, replace what we have
+function loadEphemeral() {
+    if (typeof(window.localStorage) !== "undefined") {
+        var str = window.localStorage.getItem("ponymail_config_ephemeral")
+        if (str) {
+            var eprefs = JSON.parse(str)
+            for (i in prefs) {
+                if (eprefs[i]) {
+                    prefs[i] = eprefs[i]
+                }
+            }
+        }
+        
+    }
+}
+
 function isArray(obj) {
     return (obj && obj.constructor && obj.constructor == Array)
 }
@@ -3135,7 +3160,7 @@ function buildStats(json, state, show) {
     var btn = document.createElement('a')
     btn.setAttribute("href", "javascript:void(0);")
     btn.setAttribute("class", "btn btn-warning")
-    btn.setAttribute("onclick", "prefs.hideStats='yes'; buildStats(old_json, old_state, false);")
+    btn.setAttribute("onclick", "prefs.hideStats='yes'; saveEphemeral(); buildStats(old_json, old_state, false);")
     btn.style.marginRight = "10px"
     btn.style.marginTop = "10px"
     btn.innerHTML = "Hide stats"
@@ -3148,7 +3173,7 @@ function buildStats(json, state, show) {
             document.getElementById('emails_parent').style.width = "calc(100% - 190px)"
         }
         stats.setAttribute("class", "col-md-1 vertical-text")
-        stats.innerHTML = "<div onclick=\"prefs.hideStats='no'; buildStats(old_json, old_state, true);\">Show stats panel..</div>"
+        stats.innerHTML = "<div onclick=\"prefs.hideStats='no'; saveEphemeral(); buildStats(old_json, old_state, true);\">Show stats panel..</div>"
     }
     if (prefs.hideStats == 'no' || show == true) {
         stats.setAttribute("class", "hidden-xs hidden-sm col-md-3 col-lg-3")
@@ -3173,6 +3198,7 @@ function buildStats(json, state, show) {
 
 // buildPage: build the entire page!
 function buildPage(json, state) {
+    loadEphemeral(); // load ephem config if need be
     start = new Date().getTime()
     pb_refresh = start
     json = json ? json : old_json