+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added a caching system for search results to reduce load times
 - The statistics panel is now hidden by default
+- Config changes are now persistent through reloads, even if not logged in
   
 ## CHANGES in 0.5a: